react_js:patch_package:调整update.sh，解决非yarn环境下jsview-react同步会出错问题

+++ react_js/patch_package/update.sh
@@ -2,7 +2,13 @@
 
 # 已经在package.json做关联,不需在此拷贝了
 #echo UPDATE jsview-react
-cp -a jsview-react ../node_modules/
+touch jsview-react/folder_mark.txt
+if [ ! -f "jsview-react/folder_mark.txt" ];then
+	echo UPDATE jsview-react
+	cp -a jsview-react ../node_modules/
+else
+	echo jsview-react is soft-linked to node_modules
+fi
 
 echo UDPATE react-script/config/webpack.config.js
 cp react-scripts/config/webpack.config.js ../node_modules/react-scripts/config/webpack.config.js
